update existing friend notification querries to use friend querries
