+++ server/src/schemas/schema.graphql
@@ -119,8 +119,8 @@ type Query {
   getAccountWithProfileData(email: String!): Account
   getAllFriendships: [Friendship]
   getAllNotifications: [Notification]
-  getFriendRequests: [Notification]
-  getSentFriendRequests: [Notification]
+  getRecievedFriendRequests: [Friendship]
+  getSentFriendRequests: [Friendship]
   getNewNotificationCount: Int
 }
